Change CasSecurityCondition scope to be able to use outside package
Closes #89
